astral-sh/ty#4360 astral-sh/ty#4135

+++ tests/test_groupby.py
@@ -350,6 +350,7 @@ def test_series_groupby_resample() -> None:
         DataFrame,
     )
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def f(val: Series[int] | Series[float]) -> float:
         return val.mean()
 
@@ -384,9 +385,9 @@ def test_series_groupby_resample() -> None:
     def s2series(val: Series) -> Series:
         return Series(val)
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def s2scalar(val: Series[int] | Series[float]) -> float:
-        # pyrefly: ignore[unnecessary-type-conversion]
-        return float(val.mean())
+        return val.mean()
 
     check(assert_type(GB_S.resample("ME").aggregate(np.sum), Series), Series)
     check(
@@ -577,14 +578,15 @@ def test_series_groupby_rolling() -> None:
         DataFrame,
     )
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def f(val: Series[int] | Series[float]) -> float:
         return val.mean()
 
     check(assert_type(GB_S.rolling(1).aggregate(f), Series), Series)
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def s2scalar(val: Series[int] | Series[float]) -> float:
-        # pyrefly: ignore[unnecessary-type-conversion]
-        return float(val.mean())
+        return val.mean()
 
     check(assert_type(GB_S.rolling(1).aggregate(s2scalar), Series), Series)
 
@@ -758,14 +760,15 @@ def test_series_groupby_expanding() -> None:
         DataFrame,
     )
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def f(val: Series[int] | Series[float]) -> float:
         return val.mean()
 
     check(assert_type(GB_S.expanding(1).aggregate(f), Series), Series)
 
+    # TODO: use `val: Series` instead astral-sh/ty#4360 astral-sh/ty#4135
     def s2scalar(val: Series[int] | Series[float]) -> float:
-        # pyrefly: ignore[unnecessary-type-conversion]
-        return float(val.mean())
+        return val.mean()
 
     check(assert_type(GB_S.expanding(1).aggregate(s2scalar), Series), Series)
